Improve async logger internals
The shutdown logic has been adjusted to work properly when triggered
from either the shutdown future or the sink itself. Additionally, the
inner sink will be closed when the async sink is shut down. This doesn't
matter for any of our existing inner sinks, but could in the future for
one that e.g. logs over the network.