Add BufByteSink and use for stdout
Our stdout appender previously make a separate write for each message,
and since io::stdout is line buffered would effectively be making a
syscall per message. Because stdout is inherently blocking, we can't
simply use a tokio BufWriter. Instead, this commit introduces a
BufByteSink which buffers data into a BytesMut and flushes a Bytegs
object for writing.

In future commits this will also be used in the rolling file appender
and will handle success or failure callbacks for individual messages.